Bug fixes
Updated cropArbScan to include original path length and time per line
as well as an explicit flag telling that this is a cropped file.

Scaling path length between radon and hybridvel now reports same values
as expected.

Batch processing is now stored on the first directory where it was
created. This allows to keep a single “a” file with multiple files,
even if these are spread across different directories

KNOWN ISSUES
Bug - while preparing batch processing of multiple files, sometimes the
list box is not populated and the GUI elements does not display