Added app called "curateRESfileGUI"
This app allows to browse a ALL_RES_FILE database and manually select portion[s] of the time series (all types of data are selected together, i.e. diameter and velocity together, even if the selection is made for one).  The app autosaves and also computes updates mean and standard deviation values.