Added [test version]  for extracting diameter from in plane vessels
Extraction of vessels diameter from 2D images (adapted code from Patrick Drew). Code for penetrating arteries (i.e. perpendicular to imaging plane) will be added later on. Currently the code is run from a script where all parameters need to be set. User needs to create a script for each FOV to be analyzed. Data is stored in same folder as original data. See the template script and the help of getVesselDiameter.m for more details.

NOTE: pathAnalyzeGUI has a still unfunctional (yet enabled) option to perform the LSPIV algorithm for fast line scans. It is not yet fully implemented and should not be used :-)